move everything to src
